Updating to 62.0.3170.0

+++ src/base/metrics/histogram_functions_unittest.cc
@@ -47,7 +47,7 @@ TEST(HistogramFunctionsTest, HistogramEnumeration) {
   // Verify the overflow & underflow bucket exists.
   UMA_HISTOGRAM_ENUMERATION(
       histogram, static_cast<int>(UMA_HISTOGRAM_TESTING_ENUM_THIRD) + 10,
-      UMA_HISTOGRAM_TESTING_ENUM_THIRD);
+      static_cast<int>(UMA_HISTOGRAM_TESTING_ENUM_THIRD));
   tester.ExpectBucketCount(
       histogram, static_cast<int>(UMA_HISTOGRAM_TESTING_ENUM_THIRD) + 1, 1);
   tester.ExpectTotalCount(histogram, 2);
